Add files via upload
A fixed area internal compression inlet, i.e. simply converging-diverging inlet, has a
capture-to-throat area ratio of 1.17. If the flow schematic is as shown below determine
the following:
(a) The Mach number at which the inlet will start
(b) The Mach number at the throat after starting, noting that once started Ac = A0.
(c) The stagnation pressure ratio of the inlet after starting with a shock at the throat
(d) The Mach number at which the inlet will unstart
(e) The fraction of mass spilled if the inlet were to unstart
M0 > 1 M > 1 M < 1 M = 1 M < 1
A*0 A0
Ac At = A*y
Shock Wave
Note that the fraction of air spilled by the inlet can be written as:
Fraction Spilled =
Ac 􀀀 A0
Ac
and that
A0
At
=
A0
A0
A0
Ay
=

A
A

Poy
Po0

M0
= f(M0; 
)
1
Question #2:
Consider a shrouded supersonic inlet, consisting of a shroud plate, 3.5o leading wedge,
with a total flow turning angle of 18.5o after isentropic compression, as shown below.
The length of the constant area throat in the inlet is equal to 4 channel heights.
Calculate the stagnation pressure recovery for a vehicle angle of attack of 0o and free
stream Mach numbers between 2.0 and 4.0. Note that at a later date you will be asked
to produce the recovery curves for angles of attack of 5o and 10o as well. The inlet
has been laid out such that at a design free stream Mach number of 2.65, and angle-ofattack
of 0o, the focal point of the wedge oblique shock and the isentropic compression
waves is on the cowl lip. Assume that the stagnation pressure losses in the subsonic
diffuser portion of the inlet is independent of flight conditions and equal to 0.95.
α
δ
ε
Ma
θs
Shroud
Shock
Ms
Mw
M1
σs
Shroud
Sidewall
Leading Edge
Cowl-lip
Ma - Free Stream
Ms - Post-Shroud Shock
Mw - Post-Wedge Shock
M1 - Cowl Entrance
σw
Wedge
Shock
σs - Shroud Shock Angle
α - Angle-of-Attack
σw - Wedge Shock Angle
δ - Wedge Angle
ε - Isentropic Truning Angle
θs - Shroud Sidewall
Leading Edge Angle